requirements.txt
sqlalchemy==1.4.45
bcrypt==4.0.1
python-jose==3.3.0
cryptography==38.0.4
ecdsa==0.18.0
pyasn1==0.4.8 
pycparser==2.21
cffi==1.15.1
passlib==1.7.4